backend/requirements.txt
Flask
Flask-CORS
pinecone-client
requests
beautifulsoup4
textblob
scikit-learn
numpy
sentence-transformers
google-generativeai
sentence_transformers
nltk
dotenv
